refactor: move BranchCommand class to new branch_command.py
