+++ gmatpyplus/command/__init__.py
@@ -1,1 +1,2 @@
+from .branch_command import *
 from .command import *
